+++ openfisca_core/simulations.py
@@ -27,10 +27,15 @@ class CycleError(Exception):
     pass
 
 
+class SpiralError(Exception):
+    pass
+
+
 class Simulation(object):
     """
         Represents a simulation, and handles the calculation logic
     """
+
     def __init__(
             self,
             tax_benefit_system,
@@ -50,15 +55,17 @@ class Simulation(object):
         self.create_shortcuts()
 
         # To keep track of the values (formulas and periods) being calculated to detect circular definitions.
-        # See use in formulas.py.
-        # The data structure of requested_periods_by_variable_name is: {variable_name: [period1, period2]}
-        self.requested_periods_by_variable_name = {}
-        self.max_nb_cycles = None
+        # The data structure of computation_stack is:
+        # [('variable_name', 'period1'), ('variable_name', 'period2')]
+        self.computation_stack = []
+        self.invalidated_caches = set()
 
         self.debug = False
         self.trace = False
         self.opt_out_cache = False
 
+        # controls the spirals detection; check for performance impact if > 1
+        self.max_spiral_loops = 1
         self.memory_config = None
         self._data_storage_dir = None
 
@@ -123,30 +130,40 @@ class Simulation(object):
                 self.tracer.record_calculation_end(variable.name, period, cached_array, **parameters)
             return cached_array
 
-        max_nb_cycles = parameters.get('max_nb_cycles')
-        if max_nb_cycles is not None:
-            self.max_nb_cycles = max_nb_cycles
+        array = None
 
         # First, try to run a formula
-        array = self._run_formula(variable, entity, period, max_nb_cycles)
+        try:
+            self._check_for_cycle(variable, period)
+            array = self._run_formula(variable, entity, period)
 
-        # If no result, try a base function
-        if array is None and variable.base_function:
-            array = variable.base_function(holder, period)
+            # If no result, try a base function
+            if array is None and variable.base_function:
+                array = variable.base_function(holder, period)
 
-        # If no result, use the default value
-        if array is None:
+            # If no result, use the default value and cache it
+            if array is None:
+                array = holder.default_array()
+
+            holder.put_in_cache(array, period)
+        except SpiralError:
             array = holder.default_array()
+        finally:
+            if self.trace:
+                self.tracer.record_calculation_end(variable.name, period, array, **parameters)
+            self._clean_cycle_detection_data(variable.name)
 
-        self._clean_cycle_detection_data(variable.name)
-        if max_nb_cycles is not None:
-            self.max_nb_cycles = None
-
-        holder.put_in_cache(array, period)
-        if self.trace:
-            self.tracer.record_calculation_end(variable.name, period, array, **parameters)
+        self.purge_cache_of_invalid_values()
 
         return array
+
+    def purge_cache_of_invalid_values(self):
+        # We wait for the end of calculate(), signalled by an empty stack, before purging the cache
+        if self.computation_stack:
+            return
+        for (_name, _period) in self.invalidated_caches:
+            holder = self.get_holder(_name)
+            holder.delete_arrays(_period)
 
     def calculate_add(self, variable_name, period, **parameters):
         variable = self.tax_benefit_system.get_variable(variable_name)
@@ -215,7 +232,7 @@ class Simulation(object):
             self.tracer
             )
 
-    def _run_formula(self, variable, entity, period, max_nb_cycles):
+    def _run_formula(self, variable, entity, period):
         """
             Find the ``variable`` formula for the given ``period`` if it exists, and apply it to ``entity``.
         """
@@ -229,21 +246,10 @@ class Simulation(object):
         else:
             parameters_at = self.tax_benefit_system.get_parameters_at_instant
 
-        try:
-            self._check_for_cycle(variable, period)
-            if formula.__code__.co_argcount == 2:
-                array = formula(entity, period)
-            else:
-                array = formula(entity, period, parameters_at)
-        except CycleError as error:
-            self._clean_cycle_detection_data(variable.name)
-            if max_nb_cycles is None:
-                if self.trace:
-                    self.tracer.record_calculation_abortion(variable.name, period)
-                # Re-raise until reaching the first variable called with max_nb_cycles != None in the stack.
-                raise error
-            self.max_nb_cycles = None
-            return None
+        if formula.__code__.co_argcount == 2:
+            array = formula(entity, period)
+        else:
+            array = formula(entity, period, parameters_at)
 
         self._check_formula_result(array, variable, entity, period)
         return self._cast_formula_result(array, variable)
@@ -313,47 +319,40 @@ class Simulation(object):
 
     def _check_for_cycle(self, variable, period):
         """
-        Return a boolean telling if the current variable has already been called without being allowed by
-        the parameter max_nb_cycles of the calculate method.
+        Raise an exception in the case of a circular definition, where evaluating a variable for
+        a given period loops around to evaluating the same variable/period pair. Also guards, as
+        a heuristic, against "quasicircles", where the evaluation of a variable at a period involves
+        the same variable at a different period.
         """
-        def get_error_message():
-            return "Circular definition detected on formula {}@{}. Formulas and periods involved: {}.".format(
-                variable.name,
-                period,
-                ", ".join(sorted(set(
-                    "{}@{}".format(variable_name, period2)
-                    for variable_name, periods in requested_periods_by_variable_name.items()
-                    for period2 in periods
-                    ))).encode('utf-8'),
-                )
-        requested_periods_by_variable_name = self.requested_periods_by_variable_name
-        variable_name = variable.name
-        if variable_name in requested_periods_by_variable_name:
-            # Make sure the formula doesn't call itself for the same period it is being called for.
-            # It would be a pure circular definition.
-            requested_periods = requested_periods_by_variable_name[variable_name]
-            assert period not in requested_periods and (variable.definition_period != periods.ETERNITY), get_error_message()
-            if self.max_nb_cycles is None or len(requested_periods) > self.max_nb_cycles:
-                message = get_error_message()
-                if self.max_nb_cycles is None:
-                    message += ' Hint: use "max_nb_cycles = 0" to get a default value, or "= N" to allow N cycles.'
-                raise CycleError(message)
-            else:
-                requested_periods.append(period)
-        else:
-            requested_periods_by_variable_name[variable_name] = [period]
+        previous_periods = [_period for (_name, _period) in self.computation_stack if _name == variable.name]
+        self.computation_stack.append((variable.name, str(period)))
+        if str(period) in previous_periods:
+            raise CycleError("Circular definition detected on formula {}@{}".format(variable.name, period))
+        spiral = len(previous_periods) >= self.max_spiral_loops
+        if spiral:
+            self.invalidate_spiral_variables(variable)
+            message = "Quasicircular definition detected on formula {}@{} involving {}".format(variable.name, period, self.computation_stack)
+            raise SpiralError(message, variable.name)
+
+    def invalidate_spiral_variables(self, variable):
+        # Visit the stack, from the bottom (most recent) up; we know that we'll find
+        # the variable implicated in the spiral (max_spiral_loops+1) times; we keep the
+        # intermediate values computed (to avoid impacting performance) but we mark them
+        # for deletion from the cache once the calculation ends.
+        count = 0
+        for frame in reversed(self.computation_stack):
+            self.invalidated_caches.add(frame)
+            if frame[0] == variable.name:
+                count += 1
+                if count > self.max_spiral_loops:
+                    break
 
     def _clean_cycle_detection_data(self, variable_name):
         """
         When the value of a formula have been computed, remove the period from
         requested_periods_by_variable_name[variable_name] and delete the latter if empty.
         """
-
-        requested_periods_by_variable_name = self.requested_periods_by_variable_name
-        if variable_name in requested_periods_by_variable_name:
-            requested_periods_by_variable_name[variable_name].pop()
-            if len(requested_periods_by_variable_name[variable_name]) == 0:
-                del requested_periods_by_variable_name[variable_name]
+        self.computation_stack.pop()
 
     # ----- Methods to access stored values ----- #