+++ tests/core/test_cycles.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 
-
 from openfisca_core import periods
 from openfisca_core.periods import MONTH
-from openfisca_core.simulations import CycleError
 from openfisca_core.simulation_builder import SimulationBuilder
+from openfisca_core.simulations import CycleError
 from openfisca_core.variables import Variable
 
 from openfisca_country_template import CountryTaxBenefitSystem
@@ -43,7 +42,7 @@ class variable2(Variable):
         return person('variable1', period)
 
 
-# 3 <--> 4 with a period offset, but without explicit cycle allowed
+# 3 <--> 4 with a period offset
 class variable3(Variable):
     value_type = int
     entity = Person
@@ -62,7 +61,7 @@ class variable4(Variable):
         return person('variable3', period)
 
 
-# 5 -f-> 6 with a period offset, with cycle flagged but not allowed
+# 5 -f-> 6 with a period offset
 #   <---
 class variable5(Variable):
     value_type = int
@@ -70,7 +69,7 @@ class variable5(Variable):
     definition_period = MONTH
 
     def formula(person, period):
-        variable6 = person('variable6', period.last_month, max_nb_cycles = 0)
+        variable6 = person('variable6', period.last_month)
         return 5 + variable6
 
 
@@ -84,6 +83,16 @@ class variable6(Variable):
         return 6 + variable5
 
 
+class variable7(Variable):
+    value_type = int
+    entity = Person
+    definition_period = MONTH
+
+    def formula(person, period):
+        variable5 = person('variable5', period)
+        return 7 + variable5
+
+
 # december cotisation depending on november value
 class cotisation(Variable):
     value_type = int
@@ -92,77 +101,43 @@ class cotisation(Variable):
 
     def formula(person, period):
         if period.start.month == 12:
-            return 2 * person('cotisation', period.last_month, max_nb_cycles = 1)
+            return 2 * person('cotisation', period.last_month)
         else:
             return person.empty_array() + 1
-
-
-# 7 -f-> 8 with a period offset, with explicit cycle allowed (1 level)
-#   <---
-class variable7(Variable):
-    value_type = int
-    entity = Person
-    definition_period = MONTH
-
-    def formula(person, period):
-        variable8 = person('variable8', period.last_month, max_nb_cycles = 1)
-        return 7 + variable8
-
-
-class variable8(Variable):
-    value_type = int
-    entity = Person
-    definition_period = MONTH
-
-    def formula(person, period):
-        variable7 = person('variable7', period)
-        return 8 + variable7
 
 
 # TaxBenefitSystem instance declared after formulas
 tax_benefit_system = CountryTaxBenefitSystem()
 tax_benefit_system.add_variables(variable1, variable2, variable3, variable4,
-    variable5, variable6, cotisation, variable7, variable8)
+    variable5, variable6, variable7, cotisation)
 
 
 def test_pure_cycle(simulation, reference_period):
-    with raises(AssertionError):
+    with raises(CycleError):
         simulation.calculate('variable1', period = reference_period)
 
 
-def test_cycle_time_offset(simulation, reference_period):
-    with raises(CycleError):
-        simulation.calculate('variable3', period = reference_period)
+def test_spirals_result_in_default_value(simulation, reference_period):
+    variable3 = simulation.calculate('variable3', period = reference_period)
+    assert_near(variable3, [0])
 
 
-def test_allowed_cycle(simulation, reference_period):
-    """
-    Calculate variable5 then variable6 then in the order order, to verify that the first calculated variable
-    has no effect on the result.
-    """
-    variable6 = simulation.calculate('variable6', period = reference_period)
-    variable5 = simulation.calculate('variable5', period = reference_period)
-    variable6_last_month = simulation.calculate('variable6', reference_period.last_month)
-    assert_near(variable5, [5])
-    assert_near(variable6, [11])
-    assert_near(variable6_last_month, [0])
-
-
-def test_allowed_cycle_different_order(simulation, reference_period):
+def test_spiral_heuristic(simulation, reference_period):
     variable5 = simulation.calculate('variable5', period = reference_period)
     variable6 = simulation.calculate('variable6', period = reference_period)
     variable6_last_month = simulation.calculate('variable6', reference_period.last_month)
-    assert_near(variable5, [5])
+    assert_near(variable5, [11])
     assert_near(variable6, [11])
-    assert_near(variable6_last_month, [0])
+    assert_near(variable6_last_month, [11])
+
+
+def test_spiral_cache(simulation, reference_period):
+    simulation.calculate('variable7', period = reference_period)
+    cached_variable7 = simulation.get_holder('variable7').get_array(reference_period)
+    assert cached_variable7 is not None
 
 
 def test_cotisation_1_level(simulation, reference_period):
-    cotisation = simulation.calculate('cotisation', period = reference_period.last_month)
-    assert_near(cotisation, [2])
-
-
-def test_cycle_1_level(simulation, reference_period):
-    variable7 = simulation.calculate('variable7', period = reference_period)
-    # variable8 = simulation.calculate('variable8')
-    assert_near(variable7, [22])
+    month = reference_period.last_month
+    cotisation = simulation.calculate('cotisation', period = month)
+    assert_near(cotisation, [0])